Prohibit multiple publications of the same type with the same target coordinates
When defining multiple publications, it is easy to accidentally give 2 publications the
same module coordinates. In particular, if no publication defines specific coordinates,
they all share the same default coordinates for the project.

When published, the second one published to a repository will overwrite the previous,
often resulting in an unusable published module.

There are valid use cases for configuring multiple publications, but only publishing one
of them (the "maven-publish/conditional-publishing" sample does just this). As such, we
need to validate late, only when a publication is actually published.

This commit validates each publication as it is published, failing the build with a
helpful message if a publication with the same coordinates has already been published
to the same repository. At this stage, the check is only performed on publications within
a single project.

Fixes #3835